+++ benchmark/fluid/fluid_benchmark.py
@@ -24,161 +24,7 @@ import paddle.fluid.core as core
 import paddle.fluid.profiler as profiler
 import paddle.fluid.transpiler.distribute_transpiler as distribute_transpiler
 
-BENCHMARK_MODELS = [
-    "machine_translation", "resnet", "vgg", "mnist", "stacked_dynamic_lstm"
-]
-
-
-def parse_args():
-    parser = argparse.ArgumentParser('Fluid model benchmarks.')
-    parser.add_argument(
-        '--model',
-        type=str,
-        choices=BENCHMARK_MODELS,
-        default='resnet',
-        help='The model to run benchmark with.')
-    parser.add_argument(
-        '--batch_size', type=int, default=32, help='The minibatch size.')
-    #  args related to learning rate
-    parser.add_argument(
-        '--learning_rate', type=float, default=0.001, help='The learning rate.')
-    parser.add_argument(
-        '--learning_rate_decay_method',
-        type=str,
-        default=None,
-        choices=['exponential', 'natural_exp', 'inverse_time'],
-        help='Learning rate decay method, can be exponential, natural_exp, inverse_time'
-    )
-    parser.add_argument(
-        '--learning_rate_decay_steps',
-        type=int,
-        default=100000,
-        help='Decay steps for learning rate decay method')
-    parser.add_argument(
-        '--learning_rate_decay_rate',
-        type=float,
-        default=0.5,
-        help='Decay rate for learning rate decay method')
-    #  args related to regularization
-    parser.add_argument(
-        '--weight_decay_regularizer_method',
-        type=str,
-        default=None,
-        choices=['L1', 'L2'],
-        help='Weight decay regularizer method, can be L1, L2')
-    parser.add_argument(
-        '--weight_decay_regularizer_coeff',
-        type=float,
-        default=0.1,
-        help='Weight decay regularizer coeff, 0.1 for default')
-    #  args related to gradient clipping
-    parser.add_argument(
-        '--gradient_clip_method',
-        type=str,
-        default=None,
-        choices=['Norm', 'GlobalNorm'],
-        help='Gradient clipping method, can be Norm, GlobalNorm')
-    parser.add_argument(
-        '--gradient_clip_norm',
-        type=float,
-        default=1.,
-        help='Gradient clipping norm, 1. for default')
-    #  args related to error clipping
-    parser.add_argument(
-        '--error_clip_method',
-        type=str,
-        default=None,
-        choices=['Value'],
-        help='Error clipping method, can be Value')
-    parser.add_argument(
-        '--error_clip_min',
-        type=float,
-        default=1e-6,
-        help='Error clipping min value, 1e-6 for default')
-    parser.add_argument(
-        '--error_clip_max',
-        type=float,
-        default=2e-6,
-        help='Error clipping max value, 2e-6 for default')
-    # TODO(wuyi): add "--use_fake_data" option back.
-    parser.add_argument(
-        '--skip_batch_num',
-        type=int,
-        default=5,
-        help='The first num of minibatch num to skip, for better performance test'
-    )
-    parser.add_argument(
-        '--iterations', type=int, default=80, help='The number of minibatches.')
-    parser.add_argument(
-        '--pass_num', type=int, default=100, help='The number of passes.')
-    parser.add_argument(
-        '--data_format',
-        type=str,
-        default='NCHW',
-        choices=['NCHW', 'NHWC'],
-        help='The data data_format, now only support NCHW.')
-    parser.add_argument(
-        '--device',
-        type=str,
-        default='GPU',
-        choices=['CPU', 'GPU'],
-        help='The device type.')
-    parser.add_argument(
-        '--gpus',
-        type=int,
-        default=1,
-        help='If gpus > 1, will use ParallelExecutor to run, else use Executor.')
-    parser.add_argument(
-        '--cpus',
-        type=int,
-        default=1,
-        help='If cpus > 1, will use ParallelDo to run, else use Executor.')
-    parser.add_argument(
-        '--data_set',
-        type=str,
-        default='flowers',
-        choices=['cifar10', 'flowers'],
-        help='Optional dataset for benchmark.')
-    parser.add_argument(
-        '--infer_only', action='store_true', help='If set, run forward only.')
-    parser.add_argument(
-        '--use_cprof', action='store_true', help='If set, use cProfile.')
-    parser.add_argument(
-        '--use_nvprof',
-        action='store_true',
-        help='If set, use nvprof for CUDA.')
-    parser.add_argument(
-        '--no_test',
-        action='store_true',
-        help='If set, do not test the testset during training.')
-    parser.add_argument(
-        '--memory_optimize',
-        action='store_true',
-        help='If set, optimize runtime memory before start.')
-    parser.add_argument(
-        '--use_fake_data',
-        action='store_true',
-        help='If set ommit the actual read data operators.')
-    parser.add_argument(
-        '--profile', action='store_true', help='If set, profile a few steps.')
-    parser.add_argument(
-        '--update_method',
-        type=str,
-        default='local',
-        choices=['local', 'pserver', 'nccl2'],
-        help='Choose parameter update method, can be local, pserver, nccl2.')
-    parser.add_argument(
-        '--no_split_var',
-        action='store_true',
-        default=False,
-        help='Whether split variables into blocks when update_method is pserver')
-    parser.add_argument(
-        '--async_mode',
-        action='store_true',
-        default=False,
-        help='Whether start pserver in async mode to support ASGD')
-    args = parser.parse_args()
-    return args
+from args import *
 
 
 def append_nccl2_prepare(trainer_id):
